Backend/requirements.txt
newspaper3k
firebase-admin
langdetect